# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Поддерживает создание платежей, проверку статусов\nи интеграцию с платежными шлюзами",
        "title": "Payment System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/namestnikoff/payment-system",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/payments": {
            "post": {
                "description": "Создает новый платеж в системе с валидацией суммы и валюты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Создать платеж",
                "parameters": [
                    {
                        "description": "Данные платежа",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Payment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON или неверные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/payments/status": {
            "get": {
                "description": "Возвращает информацию о статусе платежа",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Получить статус платежа",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.Payment"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd_api.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount — сумма платежа\nfloat64 = число с плавающей точкой, 64 бита точности\nДля денег в продакшене лучше использовать int64 (копейки/центы)\nПричина: float64 имеет ошибки округления (0.1 + 0.2 ≠ 0.3)\nПример: вместо 100.50 RUB хранить 10050 копеек",
                    "type": "number"
                },
                "currency": {
                    "description": "Currency — код валюты\nФормат: ISO 4217 (USD, EUR, RUB, GBP и т.д.)\n3 буквы, всегда в верхнем регистре",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "ID — уникальный идентификатор платежа\nstring = текстовая строка (как str в Python)\n`json:\"id\"` = JSON тег, указывает:\n  - При конвертации в JSON это поле будет называться \"id\" (маленькими)\n  - При парсинге JSON ключ \"id\" попадет в это поле\nВАЖНО: ID с большой буквы = публичное поле (видно из других пакетов)\n        id с маленькой = приватное (только внутри этого пакета)",
                    "type": "string"
                },
                "status": {
                    "description": "Status — статус платежа\nВозможные значения: \"pending\", \"succeeded\", \"failed\"\nВ продакшене лучше использовать enum (константы)",
                    "type": "string"
                }
            }
        }
    }
}